Refactor `thermal_control_test_helper.py` from folder `tests/platform_tests` to `tests/common/helpers` (#14588)
What is the motivation for this PR?
Some scripts use the shared functions in thermal_control_test_helper.py. To minimize cross-module dependencies, we move this helper script from feature specific folder tests/platform_tests to common folder tests/common/helpers.

How did you do it?
Some scripts use the shared functions in thermal_control_test_helper.py. To minimize cross-module dependencies, we move this helper script from feature specific folder tests/platform_tests to common folder tests/common/helpers.
