+++ tests/common/helpers/files/valid_policy.json
