+++ tests/common/helpers/minimum_table.py
