+++ tests/common/helpers/thermal_control_test_helper.py
@@ -114,7 +114,6 @@ def mocker_factory(localhost, duthosts, enum_rand_one_per_hwsku_hostname):
         mocker_object = None
 
         if 'mlnx' in platform or 'nvidia' in platform:
-            from tests.platform_tests.mellanox import mellanox_thermal_control_test_helper      # noqa F401
             mocker_type = BaseMocker.get_mocker_type(mocker_name)
             if mocker_type:
                 mocker_object = mocker_type(dut)
